Update diagram to include the idea of a turn, and define the point at which a game starts
Turns may be useful in the future for keeping score: the score should probably reflect not just the time taken to complete the game, but also how many times tiles were flipped. In other words, does completing the game in fewer turns indicate greater skill? The start of the game is also moved so that it is registered after the first tile is flipped, not when the board first appears.